# openapi.yaml
openapi: 3.0.0
info:
  title: 'LookDesk API'
  version: 1.0.0
servers:
  -
    url: 'https://api.lookdesk.ai'
    description: 'Production API'
paths:
...
  /api/documents:
    get:
      tags:
        - Documents
      summary: 'Get document list'
      operationId: 0eeb1a024876b0279b451f715fb6ed7f
      parameters:
        -
          name: lang
          in: query
          schema:
            type: string
        -
          name: category_id
          in: query
          schema:
            type: integer
        -
          name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 'Document list'
      security:
        -
          sanctum: []
    post:
      tags:
        - Documents
      summary: 'Create document with translations (each with its own file)'
      operationId: 01922b37e468863bea1b42f588c11978
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              properties:
                category_id:
                  type: integer
                  example: 1
                is_public:
                  type: boolean
                  example: true
                file:
                  description: 'Main document file (PDF/DOCX/Image)'
                  type: string
                  format: binary
                'translations[0][lang]':
                  type: string
                  example: en
                'translations[0][title]':
                  type: string
                  example: 'Contract EN'
                'translations[0][file]':
                  description: 'File for EN translation'
                  type: string
                  format: binary
                'translations[1][lang]':
                  type: string
                  example: ru
                'translations[1][title]':
                  type: string
                  example: 'Контракт RU'
                'translations[1][file]':
                  description: 'File for RU translation'
                  type: string
                  format: binary
              type: object
      responses:
        '201':
          description: Created
      security:
        -
          sanctum: []
  '/api/documents/{id}':
    get:
      tags:
        - Documents
      summary: 'Get document by ID'
      operationId: c6cc59b0640a09a31ed2a7965f93d67b
      parameters:
        -
          name: id
          in: path
          schema:
            type: integer
        -
          name: lang
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 'Document found'
        '404':
          description: 'Not found'
      security:
        -
          sanctum: []
    put:
      tags:
        - Documents
      summary: 'Update document and translations'
      operationId: ab4a342b912782e0fe834285390ff14c
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          multipart/form-data:
            schema:
              properties:
                category_id:
                  type: integer
                is_public:
                  type: boolean
                'translations[0][lang]':
                  type: string
                  example: en
                'translations[0][title]':
                  type: string
                  example: 'Contract EN'
                'translations[0][file]':
                  description: 'File for EN translation'
                  type: string
                  format: binary
                'translations[1][lang]':
                  type: string
                  example: ru
                'translations[1][title]':
                  type: string
                  example: 'Контракт RU'
                'translations[1][file]':
                  description: 'File for RU translation'
                  type: string
                  format: binary
              type: object
      responses:
        '200':
          description: Updated
      security:
        -
          sanctum: []
    delete:
      tags:
        - Documents
      summary: 'Delete document'
      operationId: 6b128368bb7ca99054b61fa94cd12162
      responses:
        '200':
          description: Deleted
      security:
        -
          sanctum: []
  '/api/documents/{id}/download':
    get:
      tags:
        - Documents
      summary: 'Download document file'
      operationId: 82e9bca1bbc2236748b6450b0502a5f0
      responses:
        '200':
          description: 'File downloaded'
      security:
        -
          sanctum: []
...
components:
...
  securitySchemes:
    sanctum:
      type: http
      bearerFormat: JWT
      scheme: bearer
...